+++ src/test/groovy/ru/vyarus/gradle/plugin/github/LegacyKitTest.groovy
@@ -68,7 +68,7 @@ class LegacyKitTest extends AbstractKitTest {
         then: "license section applied"
         def license = pom.licenses.license
         license.name.text() == "The MIT License"
-        license.url.text() == "https://raw.githubusercontent.com/$projectId/master/LICENSE"
+        license.url.text() == "https://raw.githubusercontent.com/$projectId/HEAD/LICENSE"
         license.distribution.text() == "repo"
 
         then: "site url applied"